Prevent Fixnum/Bignum deprecation warnings
Avro 1.8.2 is the latest released version. The pre-release version
(1.9.0.pre1) fixes these deprecation warnings, but we decided not to
upgrade to this version for a couple of reasons:

1. It's a pre-release, and usually it's best to wait for the final
version.
2. No one at the company knows about Avro, so it would be hard to test
whether the upgrade went OK. Plus it seems like a somewhat risky
upgrade because of uncertainty of whether there were any
backward-incompatible changes to the schema.
3. The release notes suggest that we have to upgrade to Java 8:
https://github.com/apache/avro/blob/3b6658dfcfea85e140420de207fb29d583a6df0e/CHANGES.txt#L11

Here's the commit that fixes the deprecation warnings in Avro's master
branch: b1233fd6468d4de546891cc12d19f905c55604cc

Here's a closed pull request to upgrade to the pre-release version
(there's some discussion):
https://github.com/academia-edu/academia-app/pull/12559